checks for files and directories

+++ orka/orka/orka.py
@@ -4,6 +4,8 @@
 
 """orka.orka: provides entry point main()."""
 import logging
+import os
+import subprocess
 from sys import argv
 from kamaki.clients import ClientError
 from cluster_errors_constants import *
@@ -12,7 +14,6 @@ from version import __version__
 from utils import ClusterRequest, ConnectionError, authenticate_escience, \
     get_user_clusters, custom_sort_factory, custom_date_format
 from time import sleep
-import os
 
 
 class _ArgCheck(object):
@@ -221,40 +222,72 @@ class HadoopCluster(object):
             logging.error(' You can upload files to active clusters only.')
             exit(error_fatal)
         try:
+            FNULL = open(os.devnull, 'w')
             filename = self.opts['source'].split("/")
             
-            """ Copying 
-            logging.log(SUMMARY, ' Creating temporary directory in master' )
-            os.system("ssh hduser@" + cluster['master_IP'] + " \"mkdir /home/hduser/orka-temp\"")
+            # check if file already exists in hdfs, 0: exists, 1: doesn't exist
+            file_exists = subprocess.call( "ssh hduser@" + cluster['master_IP'] 
+                            + " \"/usr/local/hadoop/bin/hdfs dfs -test -e " + self.opts['destination'] 
+                            + filename[len(filename)-1] + "\"", stderr=FNULL, shell=True)
+
+            if file_exists==0:
+                logging.log(SUMMARY, ' File already exists. Aborting upload.' )
+                exit()
+            else:
+                # size of file to be uploaded (in bytes)
+                size = os.path.getsize(self.opts['source'])
             
-            logging.log(SUMMARY, ' Copying local file to temporary directory' )
-            os.system("scp " + self.opts['source'] + " hduser@" 
-                      + cluster['master_IP'] + ":/home/hduser/orka-temp")
-            
-            logging.log(SUMMARY, ' Creating target directory to hdfs' )            
-            os.system("ssh hduser@" + cluster['master_IP'] + " \"/usr/local/hadoop/bin/hdfs dfs -mkdir " 
-                      + self.opts['destination'] + "\"")
-            logging.log(SUMMARY, ' Uploading file to hdfs' )
-            os.system("ssh hduser@" + cluster['master_IP'] 
-                      + " \"/usr/local/hadoop/bin/hdfs dfs -put /home/hduser/orka-temp/" 
-                      + filename[len(filename)-1] + " " + self.opts['destination'] + "\"")
-            logging.log(SUMMARY, ' Deleting temporary directory' )
-            os.system("ssh hduser@" + cluster['master_IP'] + " \"rm -r /home/hduser/orka-temp/\"")
-            """
-            
-            """ Streaming """
-            logging.log(SUMMARY, ' Creating target directory to hdfs (if not exists)' )
-            os.system("ssh hduser@" + cluster['master_IP'] + " \"/usr/local/hadoop/bin/hdfs dfs -mkdir " 
-                      + self.opts['destination'] + "\"")
-            
-            logging.log(SUMMARY, ' Start uploading file to hdfs' )
-            os.system("cat " + self.opts['source']  
+                # check available free space in hdfs
+                av = subprocess.check_output( "ssh hduser@" + cluster['master_IP'] 
+                            + " \"/usr/local/hadoop/bin/hdfs dfs -count -q / " + "\"", shell=True)
+                print(av)                
+                
+                
+                
+                # if file can be uploaded to hdfs
+
+
+
+
+                # check if directory exists
+                dir_exists = subprocess.call( "ssh hduser@" + cluster['master_IP'] 
+                    + " \"/usr/local/hadoop/bin/hdfs dfs -test -e " + self.opts['destination'] 
+                    + "\"", stderr=FNULL, shell=True)
+
+                if dir_exists==0:
+                    logging.log(SUMMARY, ' Directory already exists' )
+                else:
+                    logging.log(SUMMARY, ' Creating target directory to hdfs' )
+                    subprocess.call("ssh hduser@" + cluster['master_IP'] + " \"/usr/local/hadoop/bin/hdfs dfs -mkdir " 
+                      + self.opts['destination'] + "\"", stderr=FNULL, shell=True)
+                
+                """ Streaming """                
+                logging.log(SUMMARY, ' Start uploading file to hdfs' )            
+                subprocess.call("cat " + self.opts['source']  
                       + " | ssh hduser@" + cluster['master_IP'] 
                       + " /usr/local/hadoop/bin/hdfs dfs -put - " + self.opts['destination']
-                      + "/" + filename[len(filename)-1])
+                      + "/" + filename[len(filename)-1], stderr=FNULL, shell=True)
+                        
+                """ or Copying 
+                logging.log(SUMMARY, ' Creating temporary directory in master' )
+                os.system("ssh hduser@" + cluster['master_IP'] + " \"mkdir /home/hduser/orka-temp\"")
             
+                logging.log(SUMMARY, ' Copying local file to temporary directory' )
+                os.system("scp " + self.opts['source'] + " hduser@" 
+                      + cluster['master_IP'] + ":/home/hduser/orka-temp")
             
-            logging.log(SUMMARY, ' File uploaded to Hadoop filesystem' )
+                logging.log(SUMMARY, ' Creating target directory to hdfs' )            
+                os.system("ssh hduser@" + cluster['master_IP'] + " \"/usr/local/hadoop/bin/hdfs dfs -mkdir " 
+                      + self.opts['destination'] + "\"")
+                logging.log(SUMMARY, ' Uploading file to hdfs' )
+                os.system("ssh hduser@" + cluster['master_IP'] 
+                      + " \"/usr/local/hadoop/bin/hdfs dfs -put /home/hduser/orka-temp/" 
+                      + filename[len(filename)-1] + " " + self.opts['destination'] + "\"")
+                logging.log(SUMMARY, ' Deleting temporary directory' )
+                os.system("ssh hduser@" + cluster['master_IP'] + " \"rm -r /home/hduser/orka-temp/\"")
+                """
+            
+                logging.log(SUMMARY, ' File uploaded to Hadoop filesystem' )
         except Exception, e:
             logging.error(' Error:' + str(e.args[0]))
             exit(error_fatal)
